+++ apps/web/src/components/integrations/detail/context.ts
@@ -1,12 +1,11 @@
 import { type Integration } from "@deco/sdk";
-import { UseMutationResult } from "@tanstack/react-query";
 import { createContext, useContext as useContextReact } from "react";
 import { UseFormReturn } from "react-hook-form";
 
 export interface IContext {
   form: UseFormReturn<Integration>;
   integration: Integration;
-  updateIntegration: UseMutationResult<Integration, Error, Integration>;
+  onSubmit: (data: Integration) => void;
 }
 
 export const Context = createContext<IContext | null>(null);
